fix: 엔티티 수정
- Achievement: 조건 필드명 수정 -> criteria
- Group: 테이블명 수정 -> Community
- AssociateStats: id 필드 추가
