+++ memento/src/main/java/com/memento/server/domain/achievement/AchievementAssociate.java
@@ -8,7 +8,7 @@ import static lombok.AccessLevel.PROTECTED;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import com.memento.server.common.BaseEntity;
-import com.memento.server.domain.group.Associate;
+import com.memento.server.domain.community.Associate;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityListeners;
