+++ memento/src/main/java/com/memento/server/domain/community/AssociateStats.java
@@ -1,16 +1,16 @@
-package com.memento.server.domain.group;
+package com.memento.server.domain.community;
 
 import static jakarta.persistence.ConstraintMode.NO_CONSTRAINT;
 import static jakarta.persistence.FetchType.LAZY;
+import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
 import java.time.LocalDateTime;
 
-import com.memento.server.domain.member.Member;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.ForeignKey;
+import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToOne;
@@ -29,9 +29,12 @@ import lombok.NoArgsConstructor;
 public class AssociateStats {
 
 	@Id
+	@GeneratedValue(strategy = IDENTITY)
+	private Long id;
+
 	@OneToOne(fetch = LAZY)
-	@JoinColumn(name = "member_id", nullable = false, foreignKey = @ForeignKey(NO_CONSTRAINT))
-	private Member member;
+	@JoinColumn(name = "associate_id", nullable = false, foreignKey = @ForeignKey(NO_CONSTRAINT))
+	private Associate associate;
 
 	@Column(name = "consecutive_attendance_days", nullable = false, columnDefinition = "INT DEFAULT 0")
 	private int consecutiveAttendanceDays;
