+++ memento/src/main/java/com/memento/server/domain/community/Community.java
@@ -1,6 +1,6 @@
-package com.memento.server.domain.group;
+package com.memento.server.domain.community;
 
-import static jakarta.persistence.ConstraintMode.NO_CONSTRAINT;
+import static jakarta.persistence.ConstraintMode.*;
 import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
@@ -8,7 +8,6 @@ import static lombok.AccessLevel.PROTECTED;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import com.memento.server.common.BaseEntity;
-import com.memento.server.domain.achievement.Achievement;
 import com.memento.server.domain.member.Member;
 
 import jakarta.persistence.Column;
@@ -19,7 +18,6 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -27,36 +25,22 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "associates")
+@Table(name = "communities")
 @Getter
 @EntityListeners(AuditingEntityListener.class)
 @NoArgsConstructor(access = PROTECTED)
 @AllArgsConstructor
 @Builder
-public class Associate extends BaseEntity {
+public class Community extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = IDENTITY)
 	private Long id;
 
-	@Column(name = "nickname", length = 51, nullable = false)
-	private String nickname;
-
-	@Column(name = "profile_image_url", length = 255, nullable = true)
-	private String profileImageUrl;
-
-	@Column(name = "introduction", length = 255, nullable = true)
-	private String introduction;
-
-	@OneToOne(fetch = LAZY)
-	@JoinColumn(name = "achievement_id", nullable = true, foreignKey = @ForeignKey(NO_CONSTRAINT))
-	private Achievement achievement;
+	@Column(name = "name", length = 102, nullable = false)
+	private String name;
 
 	@ManyToOne(fetch = LAZY)
 	@JoinColumn(name = "member_id", nullable = false, foreignKey = @ForeignKey(NO_CONSTRAINT))
 	private Member member;
-
-	@ManyToOne(fetch = LAZY)
-	@JoinColumn(name = "group_id", nullable = false, foreignKey = @ForeignKey(NO_CONSTRAINT))
-	private Group group;
 }
