+++ memento/src/main/java/com/memento/server/domain/community/CommunityRepository.java
@@ -1,0 +1,6 @@
+package com.memento.server.domain.community;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommunityRepository extends JpaRepository<Community, Long> {
+}
