--- memento/src/main/java/com/memento/server/domain/group/GroupRepository.java
@@ -1,6 +1,0 @@
-package com.memento.server.domain.group;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GroupRepository extends JpaRepository<Group, Long> {
-}
